refacto(Campagne): extract campagneForm
