+++ ui/src/App.js
@@ -8,7 +8,7 @@ import QuestionnairesManaging from "./questionnaires/Managing";
 import QuestionnaireForm from "./questionnaires/QuestionnaireForm";
 import EditQuestionnaire from "./questionnaires/EditQuestionnaire";
 import PreviewQuestionnaire from "./questionnaires/PreviewQuestionnaire";
-import CampagneForm from "./campagnes/CampagneForm";
+import CreateCampagne from "./campagnes/CreateCampagne";
 import AnswerCampagne from "./campagnes/AnswerCampagne";
 import EditCampagne from "./campagnes/EditCampagne";
 import Login from "./users/Login";
@@ -22,7 +22,7 @@ function App() {
       <Routes>
         <Route element={<ProtectedRoute />}>
           <Route exact path="/" element={<Navigate to="/campagnes/gestion" />} />
-          <Route exact path="/campagnes/ajout" element={<CampagneForm />} />
+          <Route exact path="/campagnes/ajout" element={<CreateCampagne />} />
           <Route exact path="/campagnes/:id/edition" element={<EditCampagne />} />
           <Route exact path="/campagnes/gestion" element={<ViewCampagnes />} />
           <Route exact path="/temoignages/dashboard" element={<TemoignagesDashboard />} />
